+++ lib/src/utils/utils.dart
@@ -1,4 +1,5 @@
 export 'src/cross_file_image.dart';
 export 'src/crypto.dart';
 export 'src/debug.dart';
+export 'src/filtering_text_input_formatters.dart';
 export 'src/uuid.dart';